fix: Fix campaign save NumPy type error and add campaign suggestion feature
ISSUE: Campaign save was failing with "Error binding parameter 5 - probably unsupported type"

ROOT CAUSE:
- CoolProp PropsSI() returns NumPy scalar types (np.float64)
- These propagated through FluidProperties → metadata → AnalysisResult → database record
- SQLite parameter binding rejects NumPy types, only accepts native Python types

FIXES IMPLEMENTED:

1. Fix NumPy Type Issue (PRIMARY FIX)
   - core/fluid_properties.py: Wrap all PropsSI() calls with float() conversion
   - Converts np.float64 → float at source, preventing downstream propagation
   - Affected: density, viscosity, specific_heat, thermal_conductivity, gamma,
     speed_of_sound, compressibility_factor, molar_mass

2. Add Defensive Type Checking (SECONDARY DEFENSE)
   - core/campaign_manager_v2.py: Add sanitize_for_sqlite() helper
   - Converts NumPy types, arrays, and nested structures before SQLite binding
   - Provides safety net for any future NumPy values

3. Add Traceability (test_path Population)
   - core/campaign_manager_v2.py: Populate test_path from test_folder in metadata
   - Links database records back to folder structure
   - Enables traceability from campaign DB to source test folder

4. Add Campaign Suggestion Feature
   - core/test_metadata.py: Add parse_campaign_from_test_id() helper
     * Parses SYSTEM-CAMPAIGN-TYPE-RUN format (e.g., "RCS-C01-CF-001")
     * Returns dict with system, campaign, type, run components

   - core/campaign_manager_v2.py: Add campaign helper functions
     * suggest_campaign_name(): Creates suggested name from test_id + test_type
       Format: SYSTEM-TYPE-CAMPAIGN (e.g., "RCS-CF-C01")
     * check_campaign_exists(): Check if campaign DB file exists
     * create_campaign(): Updated to accept description parameter

   - pages/2_Single_Test_Analysis.py: Add smart campaign UI
     * Detects campaign from test_id
     * Shows suggestion: "Detected campaign: RCS-CF-C01 (matches folder structure)"
     * Offers to create campaign if doesn't exist
     * Pre-selects suggested campaign in dropdown
     * Bridges folder-based campaigns (organizational) with DB campaigns (analytical)

BENEFITS:
- Fixes immediate SQLite binding error
- Prevents future NumPy type issues
- Improves workflow: automatic campaign detection from test structure
- Adds traceability: links DB records to folder paths
- Reduces manual overhead: suggests campaign names automatically

TESTING:
- Test with CoolProp-based analysis (nitrogen, LOX, etc.)
- Test campaign suggestion with properly formatted test IDs
- Test backward compatibility with existing campaigns
- Verify test_path is populated in new records

FILES MODIFIED:
- core/fluid_properties.py (NumPy → float conversion)
- core/campaign_manager_v2.py (sanitization, test_path, suggestion helpers)
- core/test_metadata.py (parse_campaign_from_test_id)
- pages/2_Single_Test_Analysis.py (campaign suggestion UI)

+++ core/fluid_properties.py
@@ -383,12 +383,12 @@ def _get_properties_coolprop(fluid_name: str, T_K: float, P_Pa: float) -> FluidP
     )
     
     try:
-        # Core properties
-        props.density_kg_m3 = PropsSI('D', 'T', T_K, 'P', P_Pa, fluid_name)
-        props.viscosity_Pa_s = PropsSI('V', 'T', T_K, 'P', P_Pa, fluid_name)
-        props.specific_heat_J_kg_K = PropsSI('C', 'T', T_K, 'P', P_Pa, fluid_name)
-        props.thermal_conductivity_W_m_K = PropsSI('L', 'T', T_K, 'P', P_Pa, fluid_name)
-        
+        # Core properties (convert NumPy types to native Python floats for SQLite compatibility)
+        props.density_kg_m3 = float(PropsSI('D', 'T', T_K, 'P', P_Pa, fluid_name))
+        props.viscosity_Pa_s = float(PropsSI('V', 'T', T_K, 'P', P_Pa, fluid_name))
+        props.specific_heat_J_kg_K = float(PropsSI('C', 'T', T_K, 'P', P_Pa, fluid_name))
+        props.thermal_conductivity_W_m_K = float(PropsSI('L', 'T', T_K, 'P', P_Pa, fluid_name))
+
         # Phase determination
         phase_index = PropsSI('Phase', 'T', T_K, 'P', P_Pa, fluid_name)
         phase_map = {
@@ -400,18 +400,18 @@ def _get_properties_coolprop(fluid_name: str, T_K: float, P_Pa: float) -> FluidP
             6: 'two-phase',
         }
         props.phase = phase_map.get(int(phase_index), 'unknown')
-        
-        # Gas properties
+
+        # Gas properties (convert NumPy types to native Python floats)
         if props.phase in ['gas', 'supercritical', 'supercritical_gas']:
-            cp = PropsSI('C', 'T', T_K, 'P', P_Pa, fluid_name)
-            cv = PropsSI('O', 'T', T_K, 'P', P_Pa, fluid_name)  # O = Cv
+            cp = float(PropsSI('C', 'T', T_K, 'P', P_Pa, fluid_name))
+            cv = float(PropsSI('O', 'T', T_K, 'P', P_Pa, fluid_name))  # O = Cv
             if cv > 0:
-                props.gamma = cp / cv
-            props.speed_of_sound_m_s = PropsSI('A', 'T', T_K, 'P', P_Pa, fluid_name)
-            props.compressibility_factor = PropsSI('Z', 'T', T_K, 'P', P_Pa, fluid_name)
-        
-        # Molar mass (always available)
-        props.molar_mass_kg_mol = PropsSI('M', 'T', T_K, 'P', P_Pa, fluid_name)
+                props.gamma = float(cp / cv)
+            props.speed_of_sound_m_s = float(PropsSI('A', 'T', T_K, 'P', P_Pa, fluid_name))
+            props.compressibility_factor = float(PropsSI('Z', 'T', T_K, 'P', P_Pa, fluid_name))
+
+        # Molar mass (always available, convert to native Python float)
+        props.molar_mass_kg_mol = float(PropsSI('M', 'T', T_K, 'P', P_Pa, fluid_name))
         
         # CoolProp uncertainty estimates (rough)
         props.density_uncertainty = 0.001  # CoolProp typically <0.1% for common fluids
